+++ Common/src/main/java/dk/sdu/se4/common/data/GameData.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Data;
+package dk.sdu.se4.common.data;
 
 /**
  *
  * @author steff
  */
-public class World {
+public class GameData {
     
 }
