+++ Common/src/main/java/dk/sdu/se4/common/data/World.java
@@ -3,15 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Entity;
-
-import dk.sdu.se4.common.Data.GameData;
-import dk.sdu.se4.common.Data.World;
+package dk.sdu.se4.common.data;
 
 /**
  *
  * @author steff
  */
-public interface AIControllSystem {
-    void PathFinding(World world, GameData gameData);
+public class World {
+   
 }
