+++ Common/src/main/java/dk/sdu/se4/common/entity/AIControllSystem.java
@@ -3,12 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Data;
+package dk.sdu.se4.common.entity;
+
+import dk.sdu.se4.common.data.GameData;
+import dk.sdu.se4.common.data.World;
 
 /**
  *
  * @author steff
  */
-public class GameData {
-    
+public interface AIControllSystem {
+    void PathFinding(World world, GameData gameData);
 }
