+++ Common/src/main/java/dk/sdu/se4/common/entity/Entity.java
@@ -3,16 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Service;
+package dk.sdu.se4.common.entity;
 
 /**
  *
  * @author steff
  */
-public interface MapService {
-    String getid();
-    String getName();
-    int getHeight();
-    int getWeight();
+public class Entity {
     
 }
