+++ Common/src/main/java/dk/sdu/se4/common/entity/parts/EntityPart.java
@@ -3,12 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Entity;
+package dk.sdu.se4.common.entity.parts;
+
+import dk.sdu.se4.common.data.GameData;
+import dk.sdu.se4.common.data.World;
 
 /**
  *
  * @author steff
  */
-public class Entity {
-    
+public interface EntityPart {
+    void process(World world, GameData gameData);
 }
