+++ Common/src/main/java/dk/sdu/se4/common/service/GameService.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package dk.sdu.se4.common.service;
+
+/**
+ *
+ * @author steff
+ */
+public interface GameService {
+    // TO Do skal vi lige snakke om ?
+}
