+++ Common/src/main/java/dk/sdu/se4/common/service/MapService.java
@@ -3,12 +3,16 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Service;
+package dk.sdu.se4.common.service;
 
 /**
  *
  * @author steff
  */
-public interface GameService {
-    // TO Do skal vi lige snakke om ?
+public interface MapService {
+    String getid();
+    String getName();
+    int getHeight();
+    int getWeight();
+    
 }
