+++ Common/src/main/java/dk/sdu/se4/common/service/PostProcessorService.java
@@ -3,15 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Entity.Parts;
+package dk.sdu.se4.common.service;
 
-import dk.sdu.se4.common.Data.GameData;
-import dk.sdu.se4.common.Data.World;
+import dk.sdu.se4.common.data.GameData;
+import dk.sdu.se4.common.data.World;
 
 /**
  *
- * @author steff
+ * @author steffen
  */
-public interface EntityPart {
+public interface PostProcessorService {
     void process(World world, GameData gameData);
 }
