+++ Common/src/main/java/dk/sdu/se4/common/service/ProcessorService.java
@@ -3,10 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dk.sdu.se4.common.Service;
+package dk.sdu.se4.common.service;
 
-import dk.sdu.se4.common.Data.GameData;
-import dk.sdu.se4.common.Data.World;
+import dk.sdu.se4.common.data.GameData;
+import dk.sdu.se4.common.data.World;
 
 /**
  *
